[Gold II] Title: 꿈틀꿈틀 호석 애벌레 - 효율성, Time: 144 ms, Memory: 15980 KB, Score: 66 point -BaekjoonHub

+++ 백준/Gold/20181. 꿈틀꿈틀 호석 애벌레 － 효율성/꿈틀꿈틀 호석 애벌레 － 효율성.java
@@ -2,12 +2,12 @@ public class Main {
 
     public static void main(String[] args) throws Exception {
         int n = read(), k = read();
-        int[] arr = new int[n];
+        long[] arr = new long[n];
         for (int i = 0; i < n; i++)
             arr[i] = read();
-        int[] dp = new int[n + 1];
+        long[] dp = new long[n + 1];
         int l = 0, r = 0;
-        int sum = 0;
+        long sum = 0;
         while (r < n) {
             sum += arr[r];
             dp[r + 1] = dp[r];
